refactor(scheduler): extract email and slack destination inputs
Pull the email and Slack destination rows out of SchedulerFormSetupTab
into standalone SchedulerFormEmailInput and SchedulerFormSlackInput
components. Each now owns its own health/Slack-install checks and
validation state, so the setup tab no longer threads that state through.
Pure componentization — no behaviour change.
